added api integration for new arraival and created the dynamic books detail page
