+++ src/app/books/[id]/staticParams.ts
@@ -1,0 +1,8 @@
+import { getAllBooks } from "@/services/bookService";
+
+export async function generateStaticParams() {
+  const books = await getAllBooks();
+  return books.map((book) => ({
+    id: book.id,
+  }));
+}
